+++ CHANGES.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.1] -2022-04-26
+
+### Added
+- Detect split {mlang} tags and notify editor to use standard moodle form instead.
+- Default edit link added to labels on translation page.
+
+### Fixed
+- Code prechecks
+
 ## [0.9.0] - 2022-04-25
 
 ### Added